+++ CrestOptimizer/Helpers/PlayerData_AddHealth.cs
@@ -1,0 +1,14 @@
+﻿using HarmonyLib;
+
+namespace CrestOptimizer.Helpers
+{
+    [HarmonyPatch(typeof(PlayerData), "AddHealth")]
+    public static class PlayerData_AddHealth
+    {
+        [HarmonyPostfix]
+        public static void Postfix(PlayerData __instance, int amount)
+        {
+            FsmState_OnExit.masksHealed = true;
+        }
+    }
+}