+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 # Changelog
+## 1.0.0
+- New features: Beast Dash, Beast Pogo, Up Slash Size,
+Double Mirrors, Craft Bind Default
+- Minor changes to name/description text
+- Bug fix for Witch Silk Refund